learning-server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
boto3==1.35.77
python-jose[cryptography]==3.3.0
requests==2.32.3
python-dotenv==1.0.1
